% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{ancova_slopes_chrono_data}
\alias{ancova_slopes_chrono_data}
\title{ANCOVA Test of difference in slopes of OSL and C14 depth-age lines}
\usage{
ancova_slopes_chrono_data(dates_1989)
}
\description{
ANCOVA Test of difference in slopes of OSL and C14 depth-age lines
}
\examples{
\dontrun{
ancova_p_values <- ancova_slopes_chrono_data(chrono_data_slopes_plot$dates_1989)
}
}

