% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_conting}
\alias{bayes_conting}
\title{Bayesian contingency table analysis - not using this here}
\usage{
bayes_conting(lithics_table)
}
\description{
Bayesian contingency table analysis - not using this here
}
\examples{
\dontrun{
lens_bayes_tech_plot <- bayes_conting(lens_differences_tech_plot$dat_tech1)
lens_bayes_raw_plot <- bayes_conting(lens_differences_raw_plot$raw_table)
}
}

