% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_slope_difference}
\alias{bayes_slope_difference}
\title{Posterior difference in slopes during 15-20 ka only}
\usage{
bayes_slope_difference(sub_OSL, sub_C14)
}
\description{
Posterior difference in slopes during 15-20 ka only
}
\examples{
\dontrun{
bayes_slope_posteriors <- bayes_slope_difference(sed_rates_15_to_20_result$sub_OSL,
                                                 sed_rates_15_to_20_result$sub_C14)
}
}

