% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_slope_test}
\alias{bayes_slope_test}
\title{Bayesian test of difference in slopes  during 15-20 ka only}
\usage{
bayes_slope_test(x1_post, x2_post)
}
\description{
Bayesian test of difference in slopes  during 15-20 ka only
}
\examples{
\dontrun{
bayes_slope_difference_result <- bayes_slope_test(bayes_slope_difference_result$x1_post,
                                                  bayes_slope_difference_result$x2_post)
}
}

