% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_lithics_data}
\alias{clean_lithics_data}
\title{Clean the lithics raw data}
\usage{
clean_lithics_data(lithics)
}
\description{
Clean the lithics raw data
}
\examples{
\dontrun{
lithics_data_cleaned <- clean_lithics_data(lithics_data)
}
}

