% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_lens_plots}
\alias{combine_lens_plots}
\title{Combine raw materials and technology plots for upper-lens-lower}
\usage{
combine_lens_plots(tech_dat_plot, raw_dat_plot)
}
\description{
Combine raw materials and technology plots for upper-lens-lower
}
\examples{
\dontrun{
lens_tech_and_raw_plot <- combine_lens_plots(lens_differences_tech_plot$p1,lens_differences_raw_plot$p1)
}
}

