% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{get_chrono_data}
\alias{get_chrono_data}
\title{Read in the dates data (table 1 of the paper)}
\usage{
get_chrono_data()
}
\description{
Read in the dates data (table 1 of the paper)
}
\examples{
\dontrun{
chrono_data <- get_chrono_data()
}
}

