% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lithics_data}
\alias{get_lithics_data}
\title{Read in the lithics raw data}
\usage{
get_lithics_data()
}
\description{
Read in the lithics raw data
}
\examples{
\dontrun{
lithics_data <- get_lithics_data()
}
}

