% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{interpolate_chrono_data}
\alias{interpolate_chrono_data}
\title{Interpolate the chrono data}
\usage{
interpolate_chrono_data(dates_1989)
}
\description{
Interpolate the chrono data
}
\examples{
\dontrun{
chrono_data_interpolated <- interpolate_chrono_data(chrono_data_tidied)
}
}

