% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_chrono_data}
\alias{plot_chrono_data}
\title{Plot the chrono data}
\usage{
plot_chrono_data(dates_1989, oldest_depth, oldest, base_of_dense_depth, dense)
}
\description{
Plot the chrono data
}
\examples{
\dontrun{
chrono_data_plotted <- plot_chrono_data(chrono_data_tidied,
                       oldest_depth = chrono_data_interpolated$oldest_depth,
                       oldest = chrono_data_interpolated$oldest,
                       base_of_dense_depth = chrono_data_interpolated$base_of_dense_depth,
                       dense = chrono_data_interpolated$dense)
}
}

