% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_shell_data}
\alias{plot_shell_data}
\title{Plot shell data}
\usage{
plot_shell_data()
}
\description{
Plot shell data
}
\examples{
\dontrun{
shell_data_plotted <- plot_shell_data()
}
}

