% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotPost}
\alias{plotPost}
\title{plotPost}
\usage{
plotPost(paramSampleVec, credMass = 0.95, compVal = NULL,
  HDItextPlace = 0.7, ROPE = NULL, yaxt = NULL, ylab = NULL,
  xlab = NULL, cex.lab = NULL, cex = NULL, xlim = NULL, main = NULL,
  col = NULL, border = NULL, showMode = F, showCurve = F,
  breaks = NULL, ...)
}
\description{
plotPost
}
\examples{
\dontrun{
plotPost()
}
}

