% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{plots_lithics_data}
\alias{plots_lithics_data}
\title{Plots of the lithics raw data}
\usage{
plots_lithics_data(lithics, cal.date.lo)
}
\description{
Plots of the lithics raw data
}
\examples{
\dontrun{
lithics_data_plotted <- plots_lithics_data(lithics_data_cleaned, bayes_cp_result$cal.date.lo)
}
}

