% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{plots_lithics_rawmaterials}
\alias{plots_lithics_rawmaterials}
\title{Plots of the lithics raw materials}
\usage{
plots_lithics_rawmaterials(lithics, spit, cal.date.pr)
}
\description{
Plots of the lithics raw materials
}
\examples{
\dontrun{
lithics_rawmaterials_plotted <- plots_lithics_rawmaterials(lithics_data,
                                       lithics_data_plotted$spit,
                                       bayes_cp_result$cal.date.pr)
}
}

