% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{sed_rates_15_to_20}
\alias{sed_rates_15_to_20}
\title{Sedimentation rates during 15-20 ka only}
\usage{
sed_rates_15_to_20(dates_1989)
}
\description{
Sedimentation rates during 15-20 ka only
}
\examples{
\dontrun{
sed_rates_15_to_20_result <- sed_rates_15_to_20(chrono_data_slopes_plot$dates_1989)
}
}

