% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{stratiplot_lithics}
\alias{stratiplot_lithics}
\title{Stratigraphic plot of the lithic raw materials}
\usage{
stratiplot_lithics(plot2)
}
\description{
Stratigraphic plot of the lithic raw materials
}
\examples{
\dontrun{
lithics_rawmaterials_stratplotted <- stratiplot_lithics(lithics_rawmaterials_plotted$plot2)
}
}

