% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{testing_slopes_chrono_data}
\alias{testing_slopes_chrono_data}
\title{Test of difference in slopes of OSL and C14 depth-age lines}
\usage{
testing_slopes_chrono_data(dates_1989)
}
\description{
Test of difference in slopes of OSL and C14 depth-age lines
}
\examples{
\dontrun{
anova_p_value <- testing_slopes_chrono_data(chrono_data_slopes_plot$dates_1989)
}
}

