% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_chrono_data}
\alias{tidy_chrono_data}
\title{Tidy the chrono data}
\usage{
tidy_chrono_data(dates_1989)
}
\description{
Tidy the chrono data
}
\examples{
\dontrun{
chrono_data_tidied <- tidy_chrono_data(chrono_data)
}
}

